requirements.txt
FastAPI
requests
python-dotenv
uvicorn
fastapi-cors
openai
python-multipart